requirements.txt
click>=8.1.7
rich>=13.7.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
openai>=1.6.1
textual>=0.48.0
